Avoid using %zu or %zd with printf() and fprintf().
This prevents problems on systems where the system printf(3) is not
C99-compliant.  We use our own snprintf() on such systems so that
is safe.

+++ lib/util/mksiglist.c
@@ -34,16 +34,21 @@ sudo_dso_public int main(int argc, char *argv[]);
 int
 main(int argc, char *argv[])
 {
-    size_t i;
+    unsigned int i;
 
 #include "mksiglist.h"
 
+    /*
+     * For portability we must not use %zu below.
+     * This program is compiled with the host C compiler,
+     * so it cannot use any of the functions in libsudo_util.
+     */
     puts("const char *const sudo_sys_siglist[] = {");
     for (i = 0; i < nitems(sudo_sys_siglist); i++) {
 	if (sudo_sys_siglist[i] != NULL) {
 	    printf("    \"%s\",\n", sudo_sys_siglist[i]);
 	} else {
-	    printf("    \"Signal %zu\",\n", i);
+	    printf("    \"Signal %u\",\n", i);
 	}
     }
     puts("};");
